Better separate the painting logic from the Qt implementation.
This change begins to pull out the logic of painting the terminal from
the actual Qt implementation. The short term downside to this change
is that we now only support painting ASCII characters. We can solve that
later by exposing some virtual methods to handle UTF-8 strings or to use ICU.